Handle /status subresource in comperator
When an object has the /status subresource defined, the kube-apiserver
will automatically drop updates to the objects .status. If boxcutter is
input if with an object that has .status set, it will try to reconcile
the object every time as a "Recovered" action.

To fix this problem, this commit adds a check for the status subresource
and strips .status from the comparitor inputs if it is defined.